EBEGU-488 Bei der Änderung des Verantworlichen wird eine Fehlermeldung gezeigt
> Der Fehler trat auf weil die Fallnummer von der DB generiert wird. Um sicherzustellen, dass die Nummer sicher beim zuruecktransformieren im converter schon gesetzt ist, haben wir das speichern des falls in einer eigenen Transaktion ausgefuehrt. Dies hat zu einem Fehler gefuehrt wenn der Verwantwortliche Benutzer gewechselt wurde. Grund dafuer war, dass die Savetransaktion den verantwortlichen schon updated hat und dann wollte die Transaktion von FallResource abschliessen und den Verantwortlichen nochmal setzen.

> Die Fallnummer wird nun pro Mandant in der DB verwaltet und jeweils im pre persist gesetzt.

> Damit immer ein Mandant vorhanden ist habe ich ein db-unit skript gemacht

> Neu gibt es eine Mapped Superclass fuer Entities die mit einem Mandanten verknuept sind.

> Fallnummer ist nun BigInteger

+++ ebegu-rest/src/main/java/ch/dvbern/ebegu/api/dtos/JaxAntragDTO.java
@@ -29,7 +29,7 @@ public class JaxAntragDTO {
 	private String antragId = null;
 
 	@NotNull
-	private int fallNummer;
+	private long fallNummer;
 
 	@NotNull
 	private String familienName;
@@ -73,11 +73,11 @@ public class JaxAntragDTO {
 		this.antragId = antragId;
 	}
 
-	public int getFallNummer() {
+	public long getFallNummer() {
 		return fallNummer;
 	}
 
-	public void setFallNummer(int fallNummer) {
+	public void setFallNummer(long fallNummer) {
 		this.fallNummer = fallNummer;
 	}
