+++ ebegu-shared/src/main/metamodel/ch/dvbern/ebegu/entities/Fall_.java
@@ -8,7 +8,7 @@ import javax.persistence.metamodel.StaticMetamodel;
 @StaticMetamodel(Fall.class)
 public abstract class Fall_ extends ch.dvbern.ebegu.entities.AbstractEntity_ {
 
-	public static volatile SingularAttribute<Fall, Integer> fallNummer;
+	public static volatile SingularAttribute<Fall, Long> fallNummer;
 	public static volatile SingularAttribute<Fall, Integer> nextNumberKind;
 	public static volatile SingularAttribute<Fall, Benutzer> verantwortlicher;
